tools/error_codegen/pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "palio-error-codegen"
version = "0.1.0"
description = "Error catalog validation and artifact generation tooling for PalioBoard."
requires-python = ">=3.12"
dependencies = [
  "jinja2>=3.1.6,<4.0.0",
  "jsonschema>=4.25.1,<5.0.0",
  "pydantic>=2.12.5,<3.0.0",
  "pyyaml>=6.0,<7.0",
  "typer>=0.24,<1.0",
]

[project.scripts]
error-codegen = "error_codegen.cli:main"

[dependency-groups]
dev = [
  "pyright>=1.1.405,<2.0.0",
  "pytest>=8.3.0,<9.0.0",
  "ruff>=0.12.0,<0.13.0",
  "types-jsonschema>=4.26.0.20260202",
]

[tool.hatch.build.targets.wheel]
packages = ["src/error_codegen"]
include = ["src/error_codegen/templates/*.j2"]

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests/unit", "tests/integration"]
pythonpath = ["src", "."]

[tool.ruff]
src = ["src", "tests"]
target-version = "py312"

[tool.ruff.lint]
select = [
  "E",
  "F",
  "W",
  "I",
  "B",
  "UP",
  "C4",
  "SIM",
  "RUF",
  "ISC",
  "ARG",
  "ANN",
  "D",
  "TID",
]
ignore = [
  "D100",
  "D104",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]
"tests/snapshots/**/*.py" = ["E501"]

[tool.pyright]
include = ["src", "tests"]
exclude = ["**/__pycache__", ".venv", "**/node_modules", "**/.*", "tests/snapshots"]
pythonVersion = "3.12"
reportMissingTypeStubs = "warning"
typeCheckingMode = "basic"
extraPaths = ["src", "."]
venv = ".venv"
venvPath = "."
